% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{compute_juv_cohort_weights}
\alias{compute_juv_cohort_weights}
\title{Coefficients de pondération des cohortes de juvéniles voisines}
\usage{
compute_juv_cohort_weights(juv_wild, T, n_iter)
}
\arguments{
\item{juv_wild}{Liste des juvéniles sauvages par secteur, telle que retournée par \code{\link{compute_juv_wild_by_zone}}.}

\item{T}{Nombre d'années.}

\item{n_iter}{Nombre d'itérations MCMC.}
}
\value{
Une liste de 3 matrices (itérations x T) : \code{coef1}, \code{coef2}, \code{coef3}.
}
\description{
Calcule les coefficients de pondération des 3 cohortes de juvéniles
(t-4/t-5/t-6, t-3/t-4/t-5, t-2/t-3/t-4) utilisés pour répartir un retour
d'adultes observé à l'année t entre les 3 années de production juvénile
voisines (délai mer variable selon les individus). Basé sur les juvéniles
sauvages bruts par secteur (pas la cohorte déjà agrégée).
}
